New schedule def, validation, tests

+++ models/amount.go
@@ -4,7 +4,7 @@ import "github.com/hjkelly/zbbapi/common"
 
 // Amount is the central model for putting a dollar amount on anything.
 type Amount struct {
-	AmountCents int64 `json:"amount"`
+	AmountCents int `json:"amount"`
 }
 
 // GetValidated returns a sanitized copy, or an error if something isn't right.
